+++ area120/tables/apiv1alpha1/tables_client_example_test.go
@@ -81,6 +81,53 @@ func ExampleClient_ListTables() {
 	}
 }
 
+func ExampleClient_GetWorkspace() {
+	// import tablespb "google.golang.org/genproto/googleapis/area120/tables/v1alpha1"
+
+	ctx := context.Background()
+	c, err := tables.NewClient(ctx)
+	if err != nil {
+		// TODO: Handle error.
+	}
+
+	req := &tablespb.GetWorkspaceRequest{
+		// TODO: Fill request struct fields.
+	}
+	resp, err := c.GetWorkspace(ctx, req)
+	if err != nil {
+		// TODO: Handle error.
+	}
+	// TODO: Use resp.
+	_ = resp
+}
+
+func ExampleClient_ListWorkspaces() {
+	// import tablespb "google.golang.org/genproto/googleapis/area120/tables/v1alpha1"
+	// import "google.golang.org/api/iterator"
+
+	ctx := context.Background()
+	c, err := tables.NewClient(ctx)
+	if err != nil {
+		// TODO: Handle error.
+	}
+
+	req := &tablespb.ListWorkspacesRequest{
+		// TODO: Fill request struct fields.
+	}
+	it := c.ListWorkspaces(ctx, req)
+	for {
+		resp, err := it.Next()
+		if err == iterator.Done {
+			break
+		}
+		if err != nil {
+			// TODO: Handle error.
+		}
+		// TODO: Use resp.
+		_ = resp
+	}
+}
+
 func ExampleClient_GetRow() {
 	// import tablespb "google.golang.org/genproto/googleapis/area120/tables/v1alpha1"
 
@@ -219,6 +266,22 @@ func ExampleClient_DeleteRow() {
 		// TODO: Fill request struct fields.
 	}
 	err = c.DeleteRow(ctx, req)
+	if err != nil {
+		// TODO: Handle error.
+	}
+}
+
+func ExampleClient_BatchDeleteRows() {
+	ctx := context.Background()
+	c, err := tables.NewClient(ctx)
+	if err != nil {
+		// TODO: Handle error.
+	}
+
+	req := &tablespb.BatchDeleteRowsRequest{
+		// TODO: Fill request struct fields.
+	}
+	err = c.BatchDeleteRows(ctx, req)
 	if err != nil {
 		// TODO: Handle error.
 	}