+++ memcache/apiv1beta2/cloud_memcache_client_example_test.go
@@ -203,3 +203,28 @@ func ExampleCloudMemcacheClient_ApplyParameters() {
 	// TODO: Use resp.
 	_ = resp
 }
+
+func ExampleCloudMemcacheClient_ApplySoftwareUpdate() {
+	// import memcachepb "google.golang.org/genproto/googleapis/cloud/memcache/v1beta2"
+
+	ctx := context.Background()
+	c, err := memcache.NewCloudMemcacheClient(ctx)
+	if err != nil {
+		// TODO: Handle error.
+	}
+
+	req := &memcachepb.ApplySoftwareUpdateRequest{
+		// TODO: Fill request struct fields.
+	}
+	op, err := c.ApplySoftwareUpdate(ctx, req)
+	if err != nil {
+		// TODO: Handle error.
+	}
+
+	resp, err := op.Wait(ctx)
+	if err != nil {
+		// TODO: Handle error.
+	}
+	// TODO: Use resp.
+	_ = resp
+}