+++ pubsub/flow_controller_test.go
@@ -41,7 +41,7 @@ func TestFlowControllerCancel(t *testing.T) {
 	// Control: a context that is not done should always return nil.
 	go func() {
 		time.Sleep(5 * time.Millisecond)
-		fc.release(5)
+		fc.release(ctx, 5)
 	}()
 	if err := fc.acquire(context.Background(), 6); err != nil {
 		t.Errorf("got %v, expected nil", err)
@@ -79,7 +79,7 @@ func TestFlowControllerNoStarve(t *testing.T) {
 				case first <- 1:
 				default:
 				}
-				fc.release(1)
+				fc.release(ctx, 1)
 			}
 		}()
 	}
@@ -162,7 +162,7 @@ func TestFlowControllerSaturation(t *testing.T) {
 					if atomic.AddInt64(&curSize, -int64(test.acquireSize)) < 0 {
 						return errors.New("negative size")
 					}
-					fc.release(test.acquireSize)
+					fc.release(ctx, test.acquireSize)
 				}
 				return success
 			})
@@ -177,19 +177,20 @@ func TestFlowControllerSaturation(t *testing.T) {
 func TestFlowControllerTryAcquire(t *testing.T) {
 	t.Parallel()
 	fc := newFlowController(3, 10)
+	ctx := context.Background()
 
 	// Successfully tryAcquire 4 bytes.
-	if !fc.tryAcquire(4) {
+	if !fc.tryAcquire(ctx, 4) {
 		t.Error("got false, wanted true")
 	}
 
 	// Fail to tryAcquire 7 bytes.
-	if fc.tryAcquire(7) {
+	if fc.tryAcquire(ctx, 7) {
 		t.Error("got true, wanted false")
 	}
 
 	// Successfully tryAcquire 6 byte.
-	if !fc.tryAcquire(6) {
+	if !fc.tryAcquire(ctx, 6) {
 		t.Error("got false, wanted true")
 	}
 }
@@ -205,12 +206,12 @@ func TestFlowControllerUnboundedCount(t *testing.T) {
 	}
 
 	// Successfully tryAcquire 4 bytes.
-	if !fc.tryAcquire(4) {
+	if !fc.tryAcquire(ctx, 4) {
 		t.Error("got false, wanted true")
 	}
 
 	// Fail to tryAcquire 3 bytes.
-	if fc.tryAcquire(3) {
+	if fc.tryAcquire(ctx, 3) {
 		t.Error("got true, wanted false")
 	}
 }
@@ -223,9 +224,9 @@ func TestFlowControllerUnboundedCount2(t *testing.T) {
 	if err := fc.acquire(ctx, 4); err != nil {
 		t.Errorf("got %v, wanted no error", err)
 	}
-	fc.release(1)
-	fc.release(1)
-	fc.release(1)
+	fc.release(ctx, 1)
+	fc.release(ctx, 1)
+	fc.release(ctx, 1)
 	wantCount := int64(-2)
 	c := int64(fc.count())
 	if c != wantCount {
@@ -244,12 +245,12 @@ func TestFlowControllerUnboundedBytes(t *testing.T) {
 	}
 
 	// Successfully tryAcquire 4GB bytes.
-	if !fc.tryAcquire(4e9) {
+	if !fc.tryAcquire(ctx, 4e9) {
 		t.Error("got false, wanted true")
 	}
 
 	// Fail to tryAcquire a third message.
-	if fc.tryAcquire(3) {
+	if fc.tryAcquire(ctx, 3) {
 		t.Error("got true, wanted false")
 	}
 }